Minify: Move `getTar()` to `static.js`
`static.js` is the only file that uses it.
