+++ src/node/hooks/express/static.js
@@ -1,9 +1,32 @@
 'use strict';
 
+const fs = require('fs');
 const minify = require('../../utils/Minify');
+const path = require('path');
 const plugins = require('../../../static/js/pluginfw/plugin_defs');
+const settings = require('../../utils/Settings');
 const CachingMiddleware = require('../../utils/caching_middleware');
 const Yajsml = require('etherpad-yajsml');
+
+// Rewrite tar to include modules with no extensions and proper rooted paths.
+const getTar = () => {
+  const prefixLocalLibraryPath = (path) => {
+    if (path.charAt(0) === '$') {
+      return path.slice(1);
+    } else {
+      return `ep_etherpad-lite/static/js/${path}`;
+    }
+  };
+  const tarJson = fs.readFileSync(path.join(settings.root, 'src/node/utils/tar.json'), 'utf8');
+  const tar = {};
+  for (const [key, relativeFiles] of Object.entries(JSON.parse(tarJson))) {
+    const files = relativeFiles.map(prefixLocalLibraryPath);
+    tar[prefixLocalLibraryPath(key)] = files
+        .concat(files.map((p) => p.replace(/\.js$/, '')))
+        .concat(files.map((p) => `${p.replace(/\.js$/, '')}/index.js`));
+  }
+  return tar;
+};
 
 exports.expressCreateServer = (hookName, args, cb) => {
   // Cache both minified and static.
@@ -26,7 +49,7 @@ exports.expressCreateServer = (hookName, args, cb) => {
   });
 
   const StaticAssociator = Yajsml.associators.StaticAssociator;
-  const associations = Yajsml.associators.associationsForSimpleMapping(minify.getTar());
+  const associations = Yajsml.associators.associationsForSimpleMapping(getTar());
   const associator = new StaticAssociator(associations);
   jsServer.setAssociator(associator);
